+++ src/test/java/org/sonar/samples/asyncapi/checks/format/AAR011DefinedLicenseCheckTest.java
@@ -15,11 +15,36 @@ public class AAR011DefinedLicenseCheckTest extends BaseCheckTest {
         ruleName = "AAR011";
         check = new AAR011DefinedLicenseCheck();
         v2Path = getV2Path("format");
+        v3Path = getV3Path("format");
+        v31Path = getV31Path("format");
     }
 
     @Test
     public void verifyInV2() {
         verifyV2("defined-license.yaml");
+    }
+
+    
+    // ============= V3.0+ Tests =============
+
+    // --- V3 Tests ---
+    @Test
+    public void verifyV3WithLicense() {
+        verifyV3("with-license.yaml");
+    }
+    @Test
+    public void verifyV3WithoutLicense() {
+        verifyV3("without-license.yaml");
+    }
+
+    // --- V31 Tests ---
+    @Test
+    public void verifyV31WithLicense() {
+        verifyV31("with-license.yaml");
+    }
+    @Test
+    public void verifyV31WithoutLicense() {
+        verifyV31("without-license.yaml");
     }
 
     @Override