+++ src/test/java/org/sonar/samples/asyncapi/checks/format/AAR029MandatoryDescriptionCheckTest.java
@@ -15,11 +15,36 @@ public class AAR029MandatoryDescriptionCheckTest extends BaseCheckTest {
         ruleName = "AAR029";
         check = new AAR029MandatoryDescriptionCheck();
         v2Path = getV2Path("format");
+        v3Path = getV3Path("format");
+        v31Path = getV31Path("format");
     }
 
     @Test
     public void verifyInV2() {
         verifyV2("mandatory-description.yaml");
+    }
+
+    
+    // ============= V3.0+ Tests =============
+
+    // --- V3 Tests ---
+    @Test
+    public void verifyV3WithDescription() {
+        verifyV3("with-description.yaml");
+    }
+    @Test
+    public void verifyV3WithoutDescription() {
+        verifyV3("without-description.yaml");
+    }
+
+    // --- V31 Tests ---
+    @Test
+    public void verifyV31WithDescription() {
+        verifyV31("with-description.yaml");
+    }
+    @Test
+    public void verifyV31WithoutDescription() {
+        verifyV31("without-description.yaml");
     }
 
     @Override