+++ src/test/java/org/sonar/samples/asyncapi/checks/format/AAR032NumericParameterIntegrityCheckTest.java
@@ -15,12 +15,17 @@ public class AAR032NumericParameterIntegrityCheckTest extends BaseCheckTest {
         ruleName = "AAR032";
         check = new AAR032NumericParameterIntegrityCheck();
         v2Path = getV2Path("format");
+        v3Path = getV3Path("format");
+        v31Path = getV31Path("format");
     }
 
     @Test
     public void verifyInV2() {
         verifyV2("numeric-parameter-integrity.yaml");
     }
+
+    
+    // ============= V3.0+ Tests =============
 
     @Override
     public void verifyRule() {