+++ src/test/java/org/sonar/samples/asyncapi/checks/format/AAR035MessageTitleCheckTest.java
@@ -15,12 +15,17 @@ public class AAR035MessageTitleCheckTest extends BaseCheckTest {
         ruleName = "AAR035";
         check = new AAR035MessageTitleCheck();
         v2Path = getV2Path("format");
+        v3Path = getV3Path("format");
+        v31Path = getV31Path("format");
     }
 
     @Test
     public void verifyInV2() {
         verifyV2("message-title.yaml");
     }
+
+    
+    // ============= V3.0+ Tests =============
 
     @Override
     public void verifyRule() {