+++ src/test/java/org/sonar/samples/asyncapi/checks/format/AAR042MessageIdentifierCheckTest.java
@@ -15,12 +15,17 @@ public class AAR042MessageIdentifierCheckTest extends BaseCheckTest {
         ruleName = "AAR042";
         check = new AAR042MessageIdentifierCheck();
         v2Path = getV2Path("format");
+        v3Path = getV3Path("format");
+        v31Path = getV31Path("format");
     }
 
     @Test
     public void verifyInV2() {
         verifyV2("message-id.yaml");
     }
+
+    
+    // ============= V3.0+ Tests =============
 
     @Override
     public void verifyRule() {