+++ src/test/java/org/sonar/samples/asyncapi/checks/operations/AAR041ComponetChannelServerCheckTest.java
@@ -15,12 +15,26 @@ public class AAR041ComponetChannelServerCheckTest extends BaseCheckTest {
         ruleName = "AAR041";
         check = new AAR041ComponetChannelServerCheck();
         v2Path = getV2Path("operations");
+        v3Path = getV3Path("operations");
+        v31Path = getV31Path("operations");
     }
-    /*
+
     @Test
     public void verifyInV2() {
         verifyV2("componet-channel-server.yaml");
-    } */
+    }
+
+    // --- V3 Tests ---
+    @Test
+    public void verifyV3WithComponentServers() {
+        verifyV3("with-component-servers.yaml");
+    }
+
+    // --- V31 Tests ---
+    @Test
+    public void verifyV31WithComponentServers() {
+        verifyV31("with-component-servers.yaml");
+    }
 
     @Override
     public void verifyRule() {